- up to 2010.12.04
Changed files:
    GeoIP-db-City.spec -> 1.19

+++ GeoIP-db-City.spec
@@ -2,12 +2,12 @@ Summary:	GeoLite City - City database for GeoIP
 Summary(pl.UTF-8):	GeoLite City - baza danych miast dla GeoIP
 Name:		GeoIP-db-City
 # Updated every month:
-Version:	2010.11.02
+Version:	2010.12.04
 Release:	1
 License:	OPEN DATA LICENSE (see LICENSE.txt)
 Group:		Applications/Databases
 Source0:	http://geolite.maxmind.com/download/geoip/database/GeoLiteCity.dat.gz
-# Source0-md5:	58520d2741de06bb2b36797f7d224d76
+# Source0-md5:	05a5a51695b815d5effe79a7be5cc76f
 Source1:	http://www.maxmind.com/download/geoip/database/LICENSE.txt
 # Source1-md5:	a1381bd1aa0a0c91dc31b3f1e847cf4a
 URL:		http://www.maxmind.com/app/geolitecity
@@ -36,6 +36,11 @@ cp -a %{SOURCE0} .
 cp -a %{SOURCE1} .
 
 gunzip GeoLiteCity.dat.gz
+
+ver=$(stat -c '%y' GeoLiteCity.dat | awk '{print $1}' | tr - .)
+if [ "$ver" != %{version} ]; then
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
